requirements (5).txt
streamlit==1.38.0
pandas==1.5.3
numpy==1.23.5
statsmodels==0.14.0
folium==0.13.0
streamlit-folium==0.11.0
plotly
scikit-learn
fpdf2
json2
